refactor : 테스트 보완

+++ app/src/main/java/com/tgyuu/androidmajorcomponent/ui/component/colorpicker/core/HSV.kt
@@ -1,7 +1,7 @@
 package com.tgyuu.androidmajorcomponent.ui.component.colorpicker.core
 
 data class HSV(
-    val h: Int,
+    val h: Float,
     val s: Float,
     val v: Float,
 )
